Move suggestion files to common app dir
