+++ app/renderer/reducers/urlBarReducer.js
@@ -11,7 +11,7 @@ const {isURL, isPotentialPhishingUrl, getUrlFromInput} = require('../../../js/li
 const {getFrameByKey, getFrameKeyByTabId, activeFrameStatePath, frameStatePath, getActiveFrame, getFrameByTabId} = require('../../../js/state/frameStateUtil')
 const searchProviders = require('../../../js/data/searchProviders')
 const Immutable = require('immutable')
-const {generateNewSuggestionsList, generateNewSearchXHRResults} = require('../lib/suggestion')
+const {generateNewSuggestionsList, generateNewSearchXHRResults} = require('../../common/lib/suggestion')
 const {navigateSiteClickHandler} = require('../suggestionClickHandlers')
 const {getCurrentWindowId} = require('../currentWindow')
 const appStoreRenderer = require('../../../js/stores/appStoreRenderer')
